+++ src/collective/volto/formsupport/tests/test_store_action_form.py
@@ -146,6 +146,7 @@ class TestMailStore(unittest.TestCase):
             },
         )
         transaction.commit()
+
         self.assertEqual(response.status_code, 200)
         response = self.export_data()
         data = response.json()
